Move ClientCallDetails back to client
Leave ClientCallDetails wrapper as a convienence when modifying details in an
interceptor
