+++ src/grpc_interceptor/__init__.py
@@ -1,11 +1,12 @@
 """Simplified Python gRPC interceptors."""
 
-from grpc_interceptor.client import ClientInterceptor
+from grpc_interceptor.client import ClientCallDetails, ClientInterceptor
 from grpc_interceptor.exception_to_status import ExceptionToStatusInterceptor
 from grpc_interceptor.server import MethodName, parse_method_name, ServerInterceptor
 
 
 __all__ = [
+    "ClientCallDetails",
     "ClientInterceptor",
     "ExceptionToStatusInterceptor",
     "MethodName",
